move multipart functions to own file
